Move Query, Utils, and dictionary-related sources into the clp namespace.
