+++ components/core/src/clp/LogTypeDictionaryReader.hpp
@@ -1,14 +1,16 @@
-#ifndef LOGTYPEDICTIONARYREADER_HPP
-#define LOGTYPEDICTIONARYREADER_HPP
+#ifndef CLP_LOGTYPEDICTIONARYREADER_HPP
+#define CLP_LOGTYPEDICTIONARYREADER_HPP
 
-#include "Defs.h"
+#include "../Defs.h"
 #include "DictionaryReader.hpp"
 #include "LogTypeDictionaryEntry.hpp"
 
+namespace clp {
 /**
  * Class for reading logtype dictionaries from disk and performing operations on them
  */
 class LogTypeDictionaryReader
         : public DictionaryReader<logtype_dictionary_id_t, LogTypeDictionaryEntry> {};
+}  // namespace clp
 
-#endif  // LOGTYPEDICTIONARYREADER_HPP
+#endif  // CLP_LOGTYPEDICTIONARYREADER_HPP
