+++ components/core/src/clp/LogTypeDictionaryWriter.cpp
@@ -4,6 +4,7 @@
 
 using std::string;
 
+namespace clp {
 bool LogTypeDictionaryWriter::add_entry(
         LogTypeDictionaryEntry& logtype_entry,
         logtype_dictionary_id_t& logtype_id
@@ -35,3 +36,4 @@ bool LogTypeDictionaryWriter::add_entry(
     }
     return is_new_entry;
 }
+}  // namespace clp
