+++ components/core/src/clp/clg/CMakeLists.txt
@@ -1,5 +1,9 @@
 set(
         CLG_SOURCES
+        ../dictionary_utils.cpp
+        ../dictionary_utils.hpp
+        ../DictionaryEntry.hpp
+        ../DictionaryReader.hpp
         ../EncodedVariableInterpreter.cpp
         ../EncodedVariableInterpreter.hpp
         ../GlobalMetadataDB.hpp
@@ -13,6 +17,11 @@ set(
         ../Grep.hpp
         ../LogSurgeonReader.cpp
         ../LogSurgeonReader.hpp
+        ../LogTypeDictionaryEntry.cpp
+        ../LogTypeDictionaryEntry.hpp
+        ../LogTypeDictionaryReader.hpp
+        ../Query.cpp
+        ../Query.hpp
         ../streaming_archive/ArchiveMetadata.cpp
         ../streaming_archive/ArchiveMetadata.hpp
         ../streaming_archive/Constants.hpp
@@ -32,15 +41,18 @@ set(
         ../streaming_archive/writer/File.hpp
         ../streaming_archive/writer/Segment.cpp
         ../streaming_archive/writer/Segment.hpp
+        ../Utils.cpp
+        ../Utils.hpp
+        ../VariableDictionaryEntry.cpp
+        ../VariableDictionaryEntry.hpp
+        ../VariableDictionaryReader.hpp
+        ../VariableDictionaryWriter.cpp
+        ../VariableDictionaryWriter.hpp
         "${PROJECT_SOURCE_DIR}/src/BufferReader.cpp"
         "${PROJECT_SOURCE_DIR}/src/BufferReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/database_utils.cpp"
         "${PROJECT_SOURCE_DIR}/src/database_utils.hpp"
         "${PROJECT_SOURCE_DIR}/src/Defs.h"
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/ErrorCode.hpp"
         "${PROJECT_SOURCE_DIR}/src/ffi/encoding_methods.cpp"
         "${PROJECT_SOURCE_DIR}/src/ffi/encoding_methods.hpp"
@@ -57,9 +69,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.hpp"
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.inc"
         "${PROJECT_SOURCE_DIR}/src/ir/types.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/MySQLDB.cpp"
         "${PROJECT_SOURCE_DIR}/src/MySQLDB.hpp"
         "${PROJECT_SOURCE_DIR}/src/MySQLParamBindings.cpp"
@@ -72,8 +81,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/Platform.hpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.cpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Query.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Query.hpp"
         "${PROJECT_SOURCE_DIR}/src/ReaderInterface.cpp"
         "${PROJECT_SOURCE_DIR}/src/ReaderInterface.hpp"
         "${PROJECT_SOURCE_DIR}/src/spdlog_with_specializations.hpp"
@@ -100,13 +107,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/TimestampPattern.hpp"
         "${PROJECT_SOURCE_DIR}/src/TraceableException.hpp"
         "${PROJECT_SOURCE_DIR}/src/type_utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryReader.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryWriter.cpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryWriter.hpp"
         "${PROJECT_SOURCE_DIR}/src/version.hpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.cpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.hpp"
