+++ components/core/src/clp/clo/CMakeLists.txt
@@ -1,14 +1,23 @@
 set(
         CLO_SOURCES
+        ../dictionary_utils.cpp
+        ../dictionary_utils.hpp
+        ../DictionaryEntry.hpp
+        ../DictionaryReader.hpp
         ../EncodedVariableInterpreter.cpp
         ../EncodedVariableInterpreter.hpp
         ../Grep.cpp
         ../Grep.hpp
         ../LogSurgeonReader.cpp
         ../LogSurgeonReader.hpp
+        ../LogTypeDictionaryEntry.cpp
+        ../LogTypeDictionaryEntry.hpp
+        ../LogTypeDictionaryReader.hpp
         ../networking/socket_utils.cpp
         ../networking/socket_utils.hpp
         ../networking/SocketOperationFailed.hpp
+        ../Query.cpp
+        ../Query.hpp
         ../streaming_archive/ArchiveMetadata.cpp
         ../streaming_archive/ArchiveMetadata.hpp
         ../streaming_archive/Constants.hpp
@@ -28,15 +37,18 @@ set(
         ../streaming_archive/writer/File.hpp
         ../streaming_archive/writer/Segment.cpp
         ../streaming_archive/writer/Segment.hpp
+        ../Utils.cpp
+        ../Utils.hpp
+        ../VariableDictionaryEntry.cpp
+        ../VariableDictionaryEntry.hpp
+        ../VariableDictionaryReader.hpp
+        ../VariableDictionaryWriter.cpp
+        ../VariableDictionaryWriter.hpp
         "${PROJECT_SOURCE_DIR}/src/BufferReader.cpp"
         "${PROJECT_SOURCE_DIR}/src/BufferReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/database_utils.cpp"
         "${PROJECT_SOURCE_DIR}/src/database_utils.hpp"
         "${PROJECT_SOURCE_DIR}/src/Defs.h"
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/ErrorCode.hpp"
         "${PROJECT_SOURCE_DIR}/src/ffi/encoding_methods.cpp"
         "${PROJECT_SOURCE_DIR}/src/ffi/encoding_methods.hpp"
@@ -53,17 +65,12 @@ set(
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.hpp"
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.inc"
         "${PROJECT_SOURCE_DIR}/src/ir/types.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/PageAllocatedVector.hpp"
         "${PROJECT_SOURCE_DIR}/src/ParsedMessage.cpp"
         "${PROJECT_SOURCE_DIR}/src/ParsedMessage.hpp"
         "${PROJECT_SOURCE_DIR}/src/Platform.hpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.cpp"
         "${PROJECT_SOURCE_DIR}/src/Profiler.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Query.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Query.hpp"
         "${PROJECT_SOURCE_DIR}/src/ReaderInterface.cpp"
         "${PROJECT_SOURCE_DIR}/src/ReaderInterface.hpp"
         "${PROJECT_SOURCE_DIR}/src/spdlog_with_specializations.hpp"
@@ -92,13 +99,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/TimestampPattern.hpp"
         "${PROJECT_SOURCE_DIR}/src/TraceableException.hpp"
         "${PROJECT_SOURCE_DIR}/src/type_utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryReader.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryWriter.cpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryWriter.hpp"
         "${PROJECT_SOURCE_DIR}/src/version.hpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.cpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.hpp"
