+++ components/core/src/clp/clp/CommandLineArguments.cpp
@@ -8,8 +8,8 @@
 
 #include "../../Defs.h"
 #include "../../spdlog_with_specializations.hpp"
-#include "../../Utils.hpp"
 #include "../../version.hpp"
+#include "../Utils.hpp"
 
 namespace po = boost::program_options;
 using std::cerr;
