+++ components/core/src/clp/clp/compression.cpp
@@ -7,11 +7,11 @@
 #include <boost/uuid/random_generator.hpp>
 
 #include "../../spdlog_with_specializations.hpp"
-#include "../../Utils.hpp"
 #include "../GlobalMySQLMetadataDB.hpp"
 #include "../GlobalSQLiteMetadataDB.hpp"
 #include "../streaming_archive/writer/Archive.hpp"
 #include "../streaming_archive/writer/utils.hpp"
+#include "../Utils.hpp"
 #include "FileCompressor.hpp"
 #include "utils.hpp"
