+++ components/core/src/clp/clp/decompression.cpp
@@ -9,10 +9,10 @@
 #include "../../FileWriter.hpp"
 #include "../../spdlog_with_specializations.hpp"
 #include "../../TraceableException.hpp"
-#include "../../Utils.hpp"
 #include "../GlobalMySQLMetadataDB.hpp"
 #include "../GlobalSQLiteMetadataDB.hpp"
 #include "../streaming_archive/reader/Archive.hpp"
+#include "../Utils.hpp"
 #include "FileDecompressor.hpp"
 
 using std::cerr;
