+++ components/core/src/clp/dictionary_utils.hpp
@@ -1,11 +1,12 @@
-#ifndef DICTIONARY_UTILS_HPP
-#define DICTIONARY_UTILS_HPP
+#ifndef CLP_DICTIONARY_UTILS_HPP
+#define CLP_DICTIONARY_UTILS_HPP
 
 #include <string>
 
-#include "FileReader.hpp"
-#include "streaming_compression/Decompressor.hpp"
+#include "../FileReader.hpp"
+#include "../streaming_compression/Decompressor.hpp"
 
+namespace clp {
 void open_dictionary_for_reading(
         std::string const& dictionary_path,
         std::string const& segment_index_path,
@@ -19,5 +20,6 @@ void open_dictionary_for_reading(
 uint64_t read_dictionary_header(FileReader& file_reader);
 
 uint64_t read_segment_index_header(FileReader& file_reader);
+}  // namespace clp
 
-#endif  // DICTIONARY_UTILS_HPP
+#endif  // CLP_DICTIONARY_UTILS_HPP
