+++ components/core/src/clp/make_dictionaries_readable/CMakeLists.txt
@@ -1,18 +1,23 @@
 set(
         MAKE_DICTIONARIES_READABLE_SOURCES
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/dictionary_utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/DictionaryReader.hpp"
+        ../dictionary_utils.cpp
+        ../dictionary_utils.hpp
+        ../DictionaryEntry.hpp
+        ../DictionaryReader.hpp
+        ../LogTypeDictionaryEntry.cpp
+        ../LogTypeDictionaryEntry.hpp
+        ../LogTypeDictionaryReader.hpp
+        ../Utils.cpp
+        ../Utils.hpp
+        ../VariableDictionaryEntry.cpp
+        ../VariableDictionaryEntry.hpp
+        ../VariableDictionaryReader.hpp
         "${PROJECT_SOURCE_DIR}/src/FileReader.cpp"
         "${PROJECT_SOURCE_DIR}/src/FileReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/FileWriter.cpp"
         "${PROJECT_SOURCE_DIR}/src/FileWriter.hpp"
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.cpp"
         "${PROJECT_SOURCE_DIR}/src/ir/parsing.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/LogTypeDictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/ParsedMessage.cpp"
         "${PROJECT_SOURCE_DIR}/src/ParsedMessage.hpp"
         "${PROJECT_SOURCE_DIR}/src/ReaderInterface.cpp"
@@ -23,11 +28,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/streaming_compression/passthrough/Decompressor.hpp"
         "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.cpp"
         "${PROJECT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Utils.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.cpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryEntry.hpp"
-        "${PROJECT_SOURCE_DIR}/src/VariableDictionaryReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.cpp"
         "${PROJECT_SOURCE_DIR}/src/WriterInterface.hpp"
         "${PROJECT_SOURCE_DIR}/submodules/date/include/date/date.h"
