+++ components/core/src/clp/streaming_archive/reader/Archive.cpp
@@ -9,8 +9,8 @@
 #include <boost/filesystem.hpp>
 
 #include "../../../spdlog_with_specializations.hpp"
-#include "../../../Utils.hpp"
 #include "../../EncodedVariableInterpreter.hpp"
+#include "../../Utils.hpp"
 #include "../ArchiveMetadata.hpp"
 #include "../Constants.hpp"
