+++ components/core/src/clp/streaming_archive/writer/Archive.cpp
@@ -16,8 +16,8 @@
 
 #include "../../../ir/types.hpp"
 #include "../../../spdlog_with_specializations.hpp"
-#include "../../../Utils.hpp"
 #include "../../EncodedVariableInterpreter.hpp"
+#include "../../Utils.hpp"
 #include "../Constants.hpp"
 #include "utils.hpp"
